requirements.txt
# Core dependencies
python-dotenv
streamlit

# LLM and LangChain
openai
langchain
langchain-openai
langchain-core
langchain-community

# Observability tools
langsmith
langfuse
braintrust
autoevals

# Vector store and embeddings
faiss-cpu
# For GPU support, use: faiss-gpu

# PDF processing
PyPDF2

# Utilities
pandas
plotly
numpy